pyproject.toml

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "dv-flow-mgr"
version = "1.11.0"
dependencies = [
    'jq',
    'pydantic',
    'pyyaml',
    'rich',
    'tomli',
    "toposort",
    "svdep"
]
authors = [
    {name = "Matthew Ballance", email = "matt.ballance@gmail.com"},
]
description = "DV Flow Manager - Data-driven, parameterizable workflow orchestration"
license = {file = "LICENSE" }

[project.scripts]
dfm = "dv_flow.mgr.__main__:main"

[tool.setuptools.package-data]
"dv_flow.mgr" = ['share/*', 'share/*.json', 'share/references/*', 'share/prompts/*']

